docs(examples): introduce pkg/client SDK example and full SDK reference
Split the distributed-oidc-client example into two variants:

- distributed-oidc-client/ — rewritten as the recommended ~150-line SDK
  consumer using pkg/client with OIDC client-credentials, multi-endpoint
  failover, and topology refresh (down from ~480 lines)
- distributed-oidc-client-raw/ — the original hand-rolled net/http version
  retained as a wire-protocol reference and for non-Go consumers

Add docs/client-sdk.md: comprehensive SDK reference covering all four auth
modes (bearer, Basic, OIDC client credentials, mTLS via WithHTTPClient),
failover policy, topology refresh semantics with the 1s floor and seed
fallback, the full sentinel + *StatusError recipe set, command surface, and
production caveats (connection pooling, retry policy, OTel propagation,
OIDC refresh visibility).

Update __examples/README.md to list both variants with the SDK path flagged
as recommended, and register the new page in mkdocs.yml Reference navigation.
