+++ __examples/distributed-oidc-client/main.go
@@ -1,115 +1,48 @@
-// Example: connecting to a hypercache-server cluster as a backend
-// service, authenticating via OIDC client_credentials, and exercising
-// the full client API surface.
+// Example: connecting to a hypercache-server cluster using the
+// `pkg/client` SDK with OIDC client-credentials authentication.
 //
-// This is a runnable demo, not production-grade code. The goal is to
-// show:
+// This is the recommended shape for Go consumers. The SDK handles
+// auth-header injection, token refresh, endpoint failover, topology
+// refresh, content negotiation, and typed errors — everything the
+// hand-rolled raw example in distributed-oidc-client-raw/ has to do
+// by hand against net/http.
 //
-//  1. How a service obtains an OIDC access token (no human-in-the-loop)
-//     using the standard golang.org/x/oauth2/clientcredentials flow.
-//  2. How the token is automatically attached to every cache request
-//     via oauth2.NewClient's transport wrapper — there's no manual
-//     header bookkeeping.
-//  3. How GET /v1/me lets the client introspect the bound identity
-//     before doing real work (canary: "is my token actually valid
-//     against this cluster?").
-//  4. The PUT / GET / DELETE / batch surface against /v1/cache/*.
-//
-// See README.md in this directory for setup instructions (env vars,
-// IdP setup, scope mapping).
+// See ../../docs/client-sdk.md for the full SDK reference.
 package main
 
 import (
-	"bytes"
 	"context"
-	"encoding/base64"
+	"encoding/json"
+	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"net/url"
 	"os"
 	"strings"
 	"time"
 
-	"github.com/goccy/go-json"
-	"github.com/hyp3rd/ewrap"
-	"golang.org/x/oauth2"
 	"golang.org/x/oauth2/clientcredentials"
+
+	"github.com/hyp3rd/hypercache/pkg/client"
 )
 
 const (
-	defaultEndpoint = "http://localhost:8080"
-	defaultScopes   = "openid"
-	httpTimeout     = 10 * time.Second
-	putTTL          = 5 * time.Minute
-	batchTTL        = time.Minute
+	exampleKey   = "example-key"
+	exampleValue = "hello from sdk"
+	exampleTTL   = 5 * time.Minute
+
+	topologyRefresh = 30 * time.Second
+	discoveryPath   = "/.well-known/openid-configuration"
 )
 
-// Static sentinel errors — err113 forbids defining dynamic errors at
-// call sites. Wrapping these with %w lets downstream callers
-// errors.Is against each failure mode.
+// errEnvMissing is the sentinel mustEnv wraps when a required
+// variable is absent. Kept static so failure-mode tests could
+// errors.Is against it; in the example, run() surfaces the
+// wrapped error to stderr.
 var (
-	errIssuerMissing    = ewrap.New("OIDC_ISSUER is required")
-	errAudienceMissing  = ewrap.New("OIDC_AUDIENCE is required (must match the cache's HYPERCACHE_OIDC_AUDIENCE)")
-	errClientIDMissing  = ewrap.New("OIDC_CLIENT_ID is required")
-	errSecretMissing    = ewrap.New("OIDC_CLIENT_SECRET is required")
-	errDiscoveryHTTP    = ewrap.New("OIDC discovery returned non-200")
-	errDiscoveryMissing = ewrap.New("OIDC discovery doc missing token_endpoint")
-	errCacheStatus      = ewrap.New("cache returned non-2xx")
+	errEnvMissing          = errors.New("missing required env var")
+	errDiscoveryNoEndpoint = errors.New("OIDC discovery doc missing token_endpoint")
 )
-
-// envConfig is the full set of knobs the demo reads from the
-// environment. Mirroring HYPERCACHE_OIDC_* names where they overlap
-// with the server keeps operator mental-models consistent.
-type envConfig struct {
-	cacheEndpoint string // e.g. https://cache.example.com:8080
-	oidcIssuer    string // e.g. https://keycloak.example.com/realms/cache
-	oidcAudience  string // must match the server's HYPERCACHE_OIDC_AUDIENCE
-	oidcClientID  string
-	oidcSecret    string
-	oidcScopes    []string // OAuth scopes requested; e.g. "openid", "cache:read", "cache:write"
-}
-
-func loadEnv() (envConfig, error) {
-	cfg := envConfig{
-		cacheEndpoint: envOr("HYPERCACHE_ENDPOINT", defaultEndpoint),
-		oidcIssuer:    os.Getenv("OIDC_ISSUER"),
-		oidcAudience:  os.Getenv("OIDC_AUDIENCE"),
-		oidcClientID:  os.Getenv("OIDC_CLIENT_ID"),
-		oidcSecret:    os.Getenv("OIDC_CLIENT_SECRET"),
-		oidcScopes:    parseScopes(envOr("OIDC_SCOPES", defaultScopes)),
-	}
-
-	switch {
-	case cfg.oidcIssuer == "":
-		return envConfig{}, errIssuerMissing
-	case cfg.oidcAudience == "":
-		return envConfig{}, errAudienceMissing
-	case cfg.oidcClientID == "":
-		return envConfig{}, errClientIDMissing
-	case cfg.oidcSecret == "":
-		return envConfig{}, errSecretMissing
-	}
-
-	return cfg, nil
-}
-
-func envOr(name, fallback string) string {
-	if v := os.Getenv(name); v != "" {
-		return v
-	}
-
-	return fallback
-}
-
-func parseScopes(raw string) []string {
-	parts := strings.Fields(raw)
-	if len(parts) == 0 {
-		return []string{"openid"}
-	}
-
-	return parts
-}
 
 func main() {
 	err := run()
@@ -119,388 +52,121 @@ func main() {
 	}
 }
 
+// run is the testable main body. Splitting it from main() so
+// `defer c.Close()` actually executes on every error path —
+// a defer next to a log.Fatal in main() would silently skip.
 func run() error {
-	cfg, err := loadEnv()
+	issuer, err := mustEnv("OIDC_ISSUER")
 	if err != nil {
 		return err
 	}
 
-	client, err := buildClient(context.Background(), cfg)
+	clientID, err := mustEnv("OIDC_CLIENT_ID")
 	if err != nil {
 		return err
 	}
 
-	return demoFlow(context.Background(), client)
+	clientSecret, err := mustEnv("OIDC_CLIENT_SECRET")
+	if err != nil {
+		return err
+	}
+
+	audience, err := mustEnv("OIDC_AUDIENCE")
+	if err != nil {
+		return err
+	}
+
+	endpoints := strings.Fields(envOr("HYPERCACHE_ENDPOINTS", "http://localhost:8080"))
+
+	tokenURL, err := discoverTokenEndpoint(issuer)
+	if err != nil {
+		return fmt.Errorf("OIDC discovery: %w", err)
+	}
+
+	c, err := client.New(
+		endpoints,
+		client.WithOIDCClientCredentials(clientcredentials.Config{
+			ClientID:     clientID,
+			ClientSecret: clientSecret,
+			TokenURL:     tokenURL,
+			Scopes:       strings.Fields(envOr("OIDC_SCOPES", "openid")),
+			EndpointParams: url.Values{
+				"audience": {audience},
+			},
+		}),
+		client.WithTopologyRefresh(topologyRefresh),
+	)
+	if err != nil {
+		return fmt.Errorf("client.New: %w", err)
+	}
+
+	defer func() { _ = c.Close() }()
+
+	return demo(context.Background(), c)
 }
 
-// buildClient performs the one-time OIDC setup: discover the token
-// endpoint, configure the clientcredentials source, wrap an
-// http.Client whose transport auto-injects the bearer header.
-func buildClient(ctx context.Context, cfg envConfig) (*cacheClient, error) {
-	tokenURL, err := discoverTokenEndpoint(ctx, cfg.oidcIssuer)
+// demo runs the canonical introspect → set → get → delete sequence
+// against the SDK. Each step prints its result so the operator can
+// see exactly what worked.
+func demo(ctx context.Context, c *client.Client) error {
+	id, err := c.Identity(ctx)
 	if err != nil {
-		return nil, fmt.Errorf("OIDC discovery: %w", err)
+		return fmt.Errorf("identity: %w", err)
 	}
 
-	// clientcredentials handles the dance: caches tokens in memory,
-	// refreshes them before expiry, surfaces transport errors without
-	// retrying blindly. The audience request parameter is set via
-	// EndpointParams — most IdPs (Auth0, Okta, Keycloak with the
-	// audience mapper) require it for the resulting JWT's aud claim
-	// to match the cache's expectation.
-	ccCfg := &clientcredentials.Config{
-		ClientID:     cfg.oidcClientID,
-		ClientSecret: cfg.oidcSecret,
-		TokenURL:     tokenURL,
-		Scopes:       cfg.oidcScopes,
-		EndpointParams: url.Values{
-			"audience": {cfg.oidcAudience},
-		},
-		AuthStyle: oauth2.AuthStyleInParams,
-	}
+	fmt.Fprintf(os.Stdout, "authed as %s with %v\n", id.ID, id.Capabilities)
 
-	// oauth2.NewClient returns an *http.Client whose Transport
-	// auto-injects Authorization: Bearer <token> and refreshes the
-	// token transparently. This is the single most important
-	// affordance of x/oauth2 — every cache call below is a plain
-	// http.Client.Do with no header bookkeeping.
-	httpClient := oauth2.NewClient(ctx, ccCfg.TokenSource(ctx))
-
-	httpClient.Timeout = httpTimeout
-
-	return &cacheClient{
-		endpoint: strings.TrimRight(cfg.cacheEndpoint, "/"),
-		http:     httpClient,
-	}, nil
-}
-
-// demoFlow runs the canonical client API sequence: introspect, write,
-// read raw, read envelope, delete, batch write. Each step prints its
-// outcome so the operator can see exactly what worked and what didn't.
-func demoFlow(ctx context.Context, client *cacheClient) error {
-	out := os.Stdout
-
-	fmt.Fprintln(out, "== /v1/me (verify bound identity) ==")
-
-	me, err := client.me(ctx)
+	err = c.Set(ctx, exampleKey, []byte(exampleValue), exampleTTL)
 	if err != nil {
-		return fmt.Errorf("me: %w", err)
+		return fmt.Errorf("set: %w", err)
 	}
 
-	fmt.Fprintf(out, "  resolved identity: %s\n", me.ID)
-	fmt.Fprintf(out, "  granted scopes:    %v\n\n", me.Scopes)
-
-	fmt.Fprintln(out, "== PUT /v1/cache/example-key (5 min TTL) ==")
-
-	err = client.put(ctx, "example-key", []byte("hello from oidc client"), putTTL)
+	val, err := c.Get(ctx, exampleKey)
 	if err != nil {
-		return fmt.Errorf("put: %w", err)
+		return fmt.Errorf("get: %w", err)
 	}
 
-	fmt.Fprintln(out, "  stored")
-	fmt.Fprintln(out)
+	fmt.Fprintf(os.Stdout, "Get(%q) = %q\n", exampleKey, val)
 
-	fmt.Fprintln(out, "== GET /v1/cache/example-key (raw bytes) ==")
-
-	value, err := client.getRaw(ctx, "example-key")
-	if err != nil {
-		return fmt.Errorf("get raw: %w", err)
-	}
-
-	fmt.Fprintf(out, "  value: %q\n\n", value)
-
-	fmt.Fprintln(out, "== GET /v1/cache/example-key (JSON envelope) ==")
-
-	env, err := client.getEnvelope(ctx, "example-key")
-	if err != nil {
-		return fmt.Errorf("get envelope: %w", err)
-	}
-
-	fmt.Fprintf(out, "  key:      %s\n", env.Key)
-	fmt.Fprintf(out, "  version:  %d\n", env.Version)
-	fmt.Fprintf(out, "  owners:   %v\n", env.Owners)
-	fmt.Fprintf(out, "  encoding: %s\n\n", env.ValueEncoding)
-
-	fmt.Fprintln(out, "== DELETE /v1/cache/example-key ==")
-
-	err = client.delete(ctx, "example-key")
+	err = c.Delete(ctx, exampleKey)
 	if err != nil {
 		return fmt.Errorf("delete: %w", err)
 	}
 
-	fmt.Fprintln(out, "  deleted")
-	fmt.Fprintln(out)
-
-	fmt.Fprintln(out, "== batch PUT /v1/cache/batch/put (3 keys) ==")
-
-	err = client.batchPut(ctx, map[string][]byte{
-		"batch-1": []byte("one"),
-		"batch-2": []byte("two"),
-		"batch-3": []byte("three"),
-	}, batchTTL)
-	if err != nil {
-		return fmt.Errorf("batch put: %w", err)
-	}
-
-	fmt.Fprintln(out, "  stored 3 keys")
+	fmt.Fprintln(os.Stdout, "deleted")
 
 	return nil
 }
 
-// --- cache client ---
+// --- env helpers ---
 
-// cacheClient is the thin wrapper around the typed REST surface.
-// Production users would lift this into a package; the example keeps
-// it inline so all the wire shapes are visible in one file.
-type cacheClient struct {
-	endpoint string
-	http     *http.Client
+func envOr(name, fallback string) string {
+	v := os.Getenv(name)
+	if v == "" {
+		return fallback
+	}
+
+	return v
 }
 
-// meResponse mirrors the server's wire type. Duplicated rather than
-// shared because clients should depend on the JSON contract, not the
-// server's internal struct.
-type meResponse struct {
-	ID     string   `json:"id"`
-	Scopes []string `json:"scopes"`
+func mustEnv(name string) (string, error) {
+	v := os.Getenv(name)
+	if v == "" {
+		return "", fmt.Errorf("%w: %s", errEnvMissing, name)
+	}
+
+	return v, nil
 }
 
-type itemEnvelope struct {
-	Key           string   `json:"key"`
-	Value         string   `json:"value"`
-	ValueEncoding string   `json:"value_encoding"`
-	TTLMs         int64    `json:"ttl_ms,omitempty"`
-	ExpiresAt     string   `json:"expires_at,omitempty"`
-	Version       uint64   `json:"version"`
-	Origin        string   `json:"origin,omitempty"`
-	LastUpdated   string   `json:"last_updated,omitempty"`
-	Node          string   `json:"node"`
-	Owners        []string `json:"owners"`
-}
-
-// errorResponse is the canonical 4xx/5xx envelope the cache returns.
-// Surfacing the code field gives callers stable error-classification
-// without sniffing HTTP status alone.
-type errorResponse struct {
-	Code    string `json:"code"`
-	Error   string `json:"error"`
-	Details string `json:"details,omitempty"`
-}
-
-func (c *cacheClient) me(ctx context.Context) (*meResponse, error) {
-	resp, err := c.do(ctx, http.MethodGet, "/v1/me", nil, nil)
-	if err != nil {
-		return nil, err
-	}
-	defer closeBody(resp)
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, classifyResponse(resp)
-	}
-
-	var out meResponse
-
-	err = json.NewDecoder(resp.Body).Decode(&out)
-	if err != nil {
-		return nil, fmt.Errorf("decode /v1/me: %w", err)
-	}
-
-	return &out, nil
-}
-
-func (c *cacheClient) put(ctx context.Context, key string, value []byte, ttl time.Duration) error {
-	path := "/v1/cache/" + url.PathEscape(key)
-	if ttl > 0 {
-		path += "?ttl=" + ttl.String()
-	}
-
-	resp, err := c.do(ctx, http.MethodPut, path, bytes.NewReader(value), map[string]string{
-		"Content-Type": "application/octet-stream",
-	})
-	if err != nil {
-		return err
-	}
-	defer closeBody(resp)
-
-	if resp.StatusCode != http.StatusOK {
-		return classifyResponse(resp)
-	}
-
-	return nil
-}
-
-// getRaw uses the default content negotiation — no Accept header means
-// the server returns the literal value bytes. Right for callers who
-// stored bytes and want bytes back.
-func (c *cacheClient) getRaw(ctx context.Context, key string) ([]byte, error) {
-	resp, err := c.do(ctx, http.MethodGet, "/v1/cache/"+url.PathEscape(key), nil, nil)
-	if err != nil {
-		return nil, err
-	}
-	defer closeBody(resp)
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, classifyResponse(resp)
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("read body: %w", err)
-	}
-
-	return body, nil
-}
-
-// getEnvelope explicitly asks for the JSON envelope by setting
-// Accept: application/json. Right for callers that need metadata
-// (version, owners, expiry) alongside the value.
-func (c *cacheClient) getEnvelope(ctx context.Context, key string) (*itemEnvelope, error) {
-	resp, err := c.do(ctx, http.MethodGet, "/v1/cache/"+url.PathEscape(key), nil, map[string]string{
-		"Accept": "application/json",
-	})
-	if err != nil {
-		return nil, err
-	}
-	defer closeBody(resp)
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, classifyResponse(resp)
-	}
-
-	var env itemEnvelope
-
-	err = json.NewDecoder(resp.Body).Decode(&env)
-	if err != nil {
-		return nil, fmt.Errorf("decode envelope: %w", err)
-	}
-
-	return &env, nil
-}
-
-func (c *cacheClient) delete(ctx context.Context, key string) error {
-	resp, err := c.do(ctx, http.MethodDelete, "/v1/cache/"+url.PathEscape(key), nil, nil)
-	if err != nil {
-		return err
-	}
-	defer closeBody(resp)
-
-	if resp.StatusCode != http.StatusOK && resp.StatusCode != http.StatusNoContent {
-		return classifyResponse(resp)
-	}
-
-	return nil
-}
-
-// batchPutRequest mirrors the server's POST /v1/cache/batch/put body
-// shape. Values are base64-encoded for binary-safety on the JSON wire.
-type batchPutRequest struct {
-	Items []batchPutItem `json:"items"`
-}
-
-type batchPutItem struct {
-	Key   string `json:"key"`
-	Value string `json:"value"` // base64-encoded
-	TTLMs int64  `json:"ttl_ms,omitempty"`
-}
-
-func (c *cacheClient) batchPut(ctx context.Context, items map[string][]byte, ttl time.Duration) error {
-	body := batchPutRequest{Items: make([]batchPutItem, 0, len(items))}
-	for k, v := range items {
-		body.Items = append(body.Items, batchPutItem{
-			Key:   k,
-			Value: base64.StdEncoding.EncodeToString(v),
-			TTLMs: ttl.Milliseconds(),
-		})
-	}
-
-	encoded, err := json.Marshal(body)
-	if err != nil {
-		return fmt.Errorf("marshal batch put: %w", err)
-	}
-
-	resp, err := c.do(ctx, http.MethodPost, "/v1/cache/batch/put", bytes.NewReader(encoded), map[string]string{
-		"Content-Type": "application/json",
-	})
-	if err != nil {
-		return err
-	}
-	defer closeBody(resp)
-
-	if resp.StatusCode != http.StatusOK {
-		return classifyResponse(resp)
-	}
-
-	return nil
-}
-
-// do constructs an HTTP request with the given headers and dispatches
-// it via the oauth2-wrapped client. Centralizing request construction
-// keeps each verb method short and ensures every request runs through
-// the bearer-injecting transport.
-func (c *cacheClient) do(
-	ctx context.Context,
-	method, path string,
-	body io.Reader,
-	headers map[string]string,
-) (*http.Response, error) {
-	req, err := http.NewRequestWithContext(ctx, method, c.endpoint+path, body)
-	if err != nil {
-		return nil, fmt.Errorf("build %s %s: %w", method, path, err)
-	}
-
-	for k, v := range headers {
-		req.Header.Set(k, v)
-	}
-
-	resp, err := c.http.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("dispatch %s %s: %w", method, path, err)
-	}
-
-	return resp, nil
-}
-
-// closeBody discards Close errors. The body is fully drained on a
-// successful response and we don't have anything actionable to do if
-// the underlying connection's close fails (the connection is already
-// being returned to the pool / torn down).
-func closeBody(resp *http.Response) {
-	_ = resp.Body.Close()
-}
-
-// classifyResponse parses the cache's canonical error envelope and
-// returns a typed error. Falls back to the raw status when the body
-// doesn't parse — that should only happen if a load balancer returns
-// its own non-JSON 5xx ahead of the cache.
-func classifyResponse(resp *http.Response) error {
-	body, _ := io.ReadAll(resp.Body)
-
-	var envelope errorResponse
-
-	err := json.Unmarshal(body, &envelope)
-	if err != nil || envelope.Code == "" {
-		return fmt.Errorf("%w: %s: %s", errCacheStatus, resp.Status, strings.TrimSpace(string(body)))
-	}
-
-	return fmt.Errorf("%w: %s [%s]: %s", errCacheStatus, resp.Status, envelope.Code, envelope.Error)
-}
-
-// --- OIDC discovery ---
-
-type oidcDiscovery struct {
-	TokenEndpoint string `json:"token_endpoint"`
-}
-
-// discoverTokenEndpoint fetches the IdP's OIDC discovery document and
-// returns the token_endpoint URL. Spec'd at
-// https://openid.net/specs/openid-connect-discovery-1_0.html — every
-// compliant IdP serves this at /.well-known/openid-configuration.
-func discoverTokenEndpoint(ctx context.Context, issuer string) (string, error) {
+// discoverTokenEndpoint fetches the IdP's `.well-known/openid-configuration`
+// and returns its `token_endpoint`. The SDK doesn't take on OIDC discovery
+// itself — that's an IdP concern, not a cache concern — so consumer code
+// (here, this example) wires the discovery dance and hands the resolved
+// URL to clientcredentials.Config.
+func discoverTokenEndpoint(issuer string) (string, error) {
 	issuer = strings.TrimRight(issuer, "/")
 
-	discoveryURL := issuer + "/.well-known/openid-configuration"
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, discoveryURL, nil)
+	req, err := http.NewRequestWithContext(context.Background(), http.MethodGet, issuer+discoveryPath, nil)
 	if err != nil {
 		return "", fmt.Errorf("build discovery request: %w", err)
 	}
@@ -509,13 +175,16 @@ func discoverTokenEndpoint(ctx context.Context, issuer string) (string, error) {
 	if err != nil {
 		return "", fmt.Errorf("discovery request: %w", err)
 	}
-	defer closeBody(resp)
+
+	defer func() { _ = resp.Body.Close() }()
 
 	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("%w: %s", errDiscoveryHTTP, resp.Status)
+		return "", fmt.Errorf("discovery: %w: %s", errDiscoveryNoEndpoint, resp.Status)
 	}
 
-	var doc oidcDiscovery
+	var doc struct {
+		TokenEndpoint string `json:"token_endpoint"`
+	}
 
 	err = json.NewDecoder(resp.Body).Decode(&doc)
 	if err != nil {
@@ -523,7 +192,7 @@ func discoverTokenEndpoint(ctx context.Context, issuer string) (string, error) {
 	}
 
 	if doc.TokenEndpoint == "" {
-		return "", errDiscoveryMissing
+		return "", errDiscoveryNoEndpoint
 	}
 
 	return doc.TokenEndpoint, nil
